// src/app/shared/types/user/user-status.enum.ts
export enum UserStatus {
  NEW = 'NEW',
  GET_STARTED = 'GET_STARTED',
  CHOOSE_PROFESSION = 'CHOOSE_PROFESSION',
  CONFIGURE_WEBSITE = 'CONFIGURE_WEBSITE',
  PARSING_SOURCE = 'PARSING_SOURCE',
  PARSING = 'PARSING',
  PARSING_DONE = 'PARSING_DONE',
  CHOOSE_DESIGN = 'CHOOSE_DESIGN',
  DRAFT = 'DRAFT',
  PUBLISHED = 'PUBLISHED',
}